+++ lib/anyway/config.rb
@@ -107,20 +107,25 @@ module Anyway # :nodoc:
       end
     end
 
-    attr_reader :config_name, :env_prefix
+    attr_reader :config_name, :config_path, :env_prefix
 
     # Instantiate config with specified name, loads the data and applies overrides
     #
     # Example:
     #
-    #   my_config = Anyway::Config.new(name: :my_app, load: true, overrides: { some: :value })
+    #   my_config = Anyway::Config.new(
+    #    name: :my_app,
+    #    config_path: './myconfig/myconf.yml',
+    #    load: true,
+    #    overrides: { some: :value }
+    #   )
     #
-    def initialize(name: nil, load: true, overrides: {})
+    def initialize(name: nil, config_path: nil, load: true, overrides: {})
       @config_name = name || self.class.config_name
-
       raise ArgumentError, "Config name is missing" unless @config_name
 
       @env_prefix = self.class.env_prefix || @config_name
+      @config_path = resolve_config_path(config_path)
 
       self.load(overrides) if load
     end
@@ -156,14 +161,12 @@ module Anyway # :nodoc:
     end
 
     def load_from_file(config)
-      config_path = Anyway.env.fetch(env_prefix).delete('conf') ||
-                    "./config/#{config_name}.yml"
       config.deep_merge!(parse_yml(config_path) || {}) if config_path && File.file?(config_path)
       config
     end
 
     def load_from_env(config)
-      config.deep_merge!(Anyway.env.fetch(env_prefix))
+      config.deep_merge!(env_part)
       config
     end
 
@@ -187,6 +190,18 @@ module Anyway # :nodoc:
     end
 
     private
+
+    def env_part
+      Anyway.env.fetch(env_prefix)
+    end
+
+    def resolve_config_path(config_path)
+      config_path || env_part.delete('conf') || default_config_path
+    end
+
+    def default_config_path
+      "./config/#{config_name}.yml"
+    end
 
     def set_value(key, val)
       send("#{key}=", val) if respond_to?(key)